fix: 쿠폰 삭제 예외 추가
- 관리자가 쿠폰을 삭제한 후 동일한 쿠폰을 삭제하려고 할 떄 COUPON_ALREADY_DELETED 예외 발생 추가

+++ src/main/java/com/example/ssak3/domain/coupon/service/CouponService.java
@@ -79,6 +79,9 @@ public class CouponService {
         Coupon coupon = couponRepository.findById(couponId)
                 .orElseThrow(() -> new CustomException(ErrorCode.COUPON_NOT_FOUND));
 
+        if (coupon.isDeleted()) {
+            throw new CustomException(ErrorCode.COUPON_ALREADY_DELETED);
+        }
         coupon.delete();
 
         return CouponDeleteResponse.from(coupon);
